Important: fixed setting non-formlink objects in collections
Added code to make RecordGenerator distinguish if a given property lives in an array, and if so, set the property at the indicated index. This functionality already existed when the object being set was a FormLink, but I had omitted the corresponding code for non-formlink objects). Therefore, for non-formlink objects A and B, assigning A.B = "c" would work, but A[0].B="c" would fail.