Fixed asset selection issue from previous commit
Note: actual fix is in AssetSelector.cs (moving AssignmentName to end of function). Everything else is just renaming AssetPackName to AssignmentName to respect the fact that for Replacers it's the replacer name.

Issue was that when the asset assignment was passed back for a second iteration, the generatedCombination.AssetPackName (AssignmentName) would be blank, which would then cause it to be rejected by the upstream calling function