Temporary addition to OBodyWriter
This is an attempt to get the output dictionary to match the capitalization on-drive, but it doesn't seem to work (all entries have the same capitalization, which is the primary intention, but they do not respect the capitalization on-drive). Committing this but then changing to to the simpler solution suggested by Noggog of just grouping on ModKey, which has case-insensitivity in its comparison override.